Split git-gh-command.ts to satisfy the 300-line file limit #500
