+++ package.json
@@ -1,5 +1,5 @@
 {
-	"version": "0.225.0",
+	"version": "0.226.0",
 	"name": "@joshuafolkken/kit",
 	"type": "module",
 	"license": "MIT",
